clang-tidy

+++ foo_spider_monkey_panel/js_objects/gdi_utils.h
@@ -32,8 +32,8 @@ public:
 
 public:
     JSObject* CreateImage( uint32_t w, uint32_t h );
-    JSObject* Font( const std::wstring& fontName, float pxSize, uint32_t style = 0 );
-    JSObject* FontWithOpt( size_t optArgCount, const std::wstring& fontName, float pxSize, uint32_t style );
+    JSObject* Font( const std::wstring& fontName, uint32_t pxSize, uint32_t style = 0 );
+    JSObject* FontWithOpt( size_t optArgCount, const std::wstring& fontName, uint32_t pxSize, uint32_t style );
     JSObject* Image( const std::wstring& path );
     std::uint32_t LoadImageAsync( uint32_t hWnd, const std::wstring& path );
     JSObject* LoadImageAsyncV2( uint32_t hWnd, const std::wstring& path );
